requirements.txt
Flask
openpyxl
discord.py
